+++ notifier.py
@@ -4,6 +4,7 @@ import random
 import github3
 from jinja2 import Environment, PackageLoader, select_autoescape
 import parsenvy
+from raven import Client
 import redis
 from rq import Queue
 from sqlalchemy import or_
@@ -13,9 +14,12 @@ from helpers import send_email
 
 
 REDIS_URL = parsenvy.str('REDIS_URL')
+SENTRY_DSN = parsenvy.str('SENTRY_DSN')
 
 
 q = Queue(connection=redis.from_url(REDIS_URL))
+
+client = Client(SENTRY_DSN)
 
 env = Environment(
     loader=PackageLoader('starminder', 'templates'),
@@ -30,6 +34,14 @@ users = (db.session.query(User).filter(User.time == hour,
                                .all())
 
 for user in users:
+
+    if user.email is None:
+        error_message = 'User {0} ({1}) has no email address.'.format(
+            user.github_username,
+            user.id
+        )
+        client.client.captureMessage(error_message)
+
     gh = github3.login(token=user.github_token)
     gh_user = gh.user()
     all_stars = list(gh_user.iter_starred())